provide a overall summary along with call-wise
```
Duration: 1m3s ▱▱▱
RPM    :  1168103.3
RX Rate:↑ 2.5 TiB/m
TX Rate:↓ 2.4 TiB/m
-------------
...
```
