+++ cmd/support-top-api.go
@@ -103,27 +103,27 @@ func mainSupportTopAPI(ctx *cli.Context) error {
 	// Start listening on all trace activity.
 	traceCh := client.ServiceTrace(ctxt, opts)
 
-	p := tea.NewProgram(initTraceUI())
+	filteredTraces := make(chan madmin.ServiceTraceInfo, 1)
+	ui := tea.NewProgram(initTraceStatsUI(30, filteredTraces))
+	var te error
 	go func() {
-		for apiCallInfo := range traceCh {
-			if apiCallInfo.Err != nil {
-				fatalIf(probe.NewError(apiCallInfo.Err), "Unable to fetch top API events")
+		for t := range traceCh {
+			if t.Err != nil {
+				te = t.Err
+				ui.Kill()
+				return
 			}
-			if mopts.matches(apiCallInfo) {
-				p.Send(topAPIResult{
-					apiCallInfo: apiCallInfo,
-				})
+			if mopts.matches(t) {
+				filteredTraces <- t
 			}
-			p.Send(topAPIResult{
-				apiCallInfo: madmin.ServiceTraceInfo{},
-			})
 		}
 	}()
-
-	if _, e := p.Run(); e != nil {
+	if _, e := ui.Run(); e != nil {
 		cancel()
-		fatalIf(probe.NewError(e).Trace(aliasedURL), "Unable to fetch top API events")
+		if te != nil {
+			e = te
+		}
+		fatalIf(probe.NewError(e).Trace(aliasedURL), "Unable to fetch http trace statistics")
 	}
-
 	return nil
 }
